+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 #### next release (8.7.2)
 
-- [The next improvement]
+- Add NumberParameterEditor to enable WPS AllowedValues Ranges to be set and use DefaultValue
 
 #### 8.7.1 - 2024-04-16